// example4/oneteam/src/app/store/project/project.types.ts
interface IProjectsRequestListRow {
  id: string;
  prj_code: string;
  subPrjCode: string;
  mnf_ext_prj_code?: string;
  type: string;
  name: string;
  op_code: string;
  op_name: string;
  RP_name: string;
  RP_phone: string;
  obj_addr: string;
  stage_name: string;
  mnf_prj_statusDesc: string;
  mnf_dbeg: string;
  mnf_dend: string;
  mnf_planTO_sum: string;
  avtor_fio: string;
  avtor_pos: string;
  avtor_phone: string;
  avtor_email: string;
  mnf_comp_mnf: string;
  sub_prj_statusDesc: string;
}

export interface IProjectsListRow {
  id: string;
  prj_code: string;
  subPrjCode: string;
  mnf_ext_prj_code: string;
  mnf_prj_status: string;
  mnf_prj_statusDesc: string;
  mnf_ext_status: string;
  name: string;
  op_code: string;
  op_name: string;
  RP_name: string;
  RP_phone: string;
  obj_addr: string;
  stage_name: string;
  mnf_dbeg: string;
  mnf_dend: string;
  mnf_planTO_sum: string;
  mnf_comp_mnf: string;
  sub_prj_statusDesc: string;
}
// ...
export const ProjectsComputedProperty = {
  projects: 'projects',
  projectRequests: 'projectRequests',
} as const;

export interface IProjectsListDataForTable
  extends Pick<IProjectsRequestListRow, 'id' | 'type' | 'name' | 'op_name' | 'obj_addr'> {
  reg_code: string;
  reg_num: string;
  rp_data: string;
  stage: string;
  period: string;
  plan_sum: string;
  comp: string;
}

export interface IProjectsDataForTable extends Pick<IProjectsListRow, 'id' | 'name' | 'op_name' | 'obj_addr'> {
  reg_code: string;
  reg_num: string;
  status: string;
  rp_data: string;
  stage: string;
  period: string;
  plan_sum: string;
  comp: string;
}